Candidate : Create Candidate info status update api

+++ routes/web.php
@@ -18,7 +18,10 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('welcome');
 });
+
 Route::get('/migration', function () {
-    Artisan::call('migrate:fresh', ['--seed'=>true]);
-    echo "Migration and seed done successfully";
+    if(env('APP_ENV') === 'local'){
+        Artisan::call('migrate:fresh', ['--seed'=>true]);
+        echo "Migration and seed done successfully";
+    }
 });
